Made minor modifications. Set up views and routes. for some endpoints.

+++ app/controllers/mobile_controller.rb
@@ -1,12 +1,4 @@
 class MobileController < ApplicationController
-  def authenticate
-    @user = User.find_by(email: params[:email].downcase)
-    if @user && @user.authenticate(params[:password])
-      return @user
-    else
-      return false
-    end
-  end
   
   def login
     if authenticate
@@ -38,12 +30,12 @@ class MobileController < ApplicationController
        .offset(params[:page_size].to_i * (params[:page].to_i - 1))
        .order("#{params[:order]} #{params[:direction]}")
        .all.each
-    render :json => @items
+    render :json => @items.to_json(:include => :subitem)
   end
   
   def item
     if @item = Item.find_by(id: params[:item])
-      render :json => @item
+      render :json => @item.to_json(:include => :subitem)
     else
       render :json => {error: "Item not found"}
     end
@@ -60,17 +52,29 @@ class MobileController < ApplicationController
   end
   
   def purchase
-    session[:cart].each do |item_id, quantity_bought|
-        item = Item.find_by(id: item_id)
-        # Generates receipt
-        get_receipt(user, item.user, item, quantity_bought, quantity_bought * item.price)
-        quantity_remaining = item.quantity - quantity_bought
-        if quantity_remaining == 0
-          item.destroy
-        else
-          item.update_attribute(:quantity, quantity_remaining)
+    if authenticate
+      params[:items].each do |item_id, quantity_bought|
+          item = Item.find_by(id: item_id)
+          # Generates receipt
+          get_receipt(user, item.user, item, quantity_bought, quantity_bought * item.price)
+          quantity_remaining = item.quantity - quantity_bought
+          if quantity_remaining == 0
+            item.destroy
+          else
+            item.update_attribute(:quantity, quantity_remaining)
+          end
         end
-      end
-      clear_cart
+        clear_cart
+    end
+  end
+  
+  private
+  def authenticate
+    @user = User.find_by(email: params[:email].downcase)
+    if @user && @user.authenticate(params[:password])
+      return @user
+    else
+      return false
+    end
   end
 end
